perf: avoid recreating plugins twice
